Bug fixes to collecting valid regions and respecting alignment
Former-commit-id: 924fbc45f0b7fd47be5da2f6326b8f73eea9ffd2

+++ Anathema/GUI/Tools/Scanners/GUIChangeCounter.cs
@@ -38,19 +38,19 @@ namespace Anathema
 
         private void SetMinChanges()
         {
-            Int32 MinChanges = MinChangesTrackBar.Value;
+            UInt16 MinChanges = (UInt16)MinChangesTrackBar.Value;
             MinChangesValueLabel.Text = MinChanges.ToString();
             ChangeCounterPresenter.SetMinChanges(MinChanges);
         }
 
         private void SetMaxChanges()
         {
-            Int32 MaxChanges = MaxChangesTrackBar.Value;
+            UInt16 MaxChanges = (UInt16)MaxChangesTrackBar.Value;
             String MaxChangesString = MaxChanges.ToString();
             
             if (MaxChanges == MaxChangesTrackBar.Maximum)
             {
-                MaxChanges = Int32.MaxValue;
+                MaxChanges = UInt16.MaxValue;
                 MaxChangesString = "Inf";
             }
